Moved backport of `XCTUnwrap` into dedicated file
